Move abstract classes into the base package
