+++ ehri-ws/src/main/java/eu/ehri/extension/PermissionsResource.java
@@ -19,6 +19,7 @@
 
 package eu.ehri.extension;
 
+import eu.ehri.extension.base.AbstractRestResource;
 import eu.ehri.project.acl.*;
 import eu.ehri.project.core.Tx;
 import eu.ehri.project.definitions.Entities;
