+++ ehri-ws/src/main/java/eu/ehri/extension/base/AbstractRestResource.java
@@ -17,7 +17,7 @@
  * permissions and limitations under the Licence.
  */
 
-package eu.ehri.extension;
+package eu.ehri.extension.base;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
