+++ ehri-ws/src/test/java/eu/ehri/extension/test/AccessRestClientTest.java
@@ -22,7 +22,7 @@ package eu.ehri.extension.test;
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
 import com.sun.jersey.core.util.MultivaluedMapImpl;
-import eu.ehri.extension.AbstractRestResource;
+import eu.ehri.extension.base.AbstractRestResource;
 import eu.ehri.extension.AccessResource;
 import eu.ehri.project.definitions.Entities;
 import org.junit.Test;
