+++ ehri-ws/src/test/java/eu/ehri/extension/test/AnnotationRestClientTest.java
@@ -21,7 +21,7 @@ package eu.ehri.extension.test;
 
 import com.sun.jersey.api.client.ClientResponse;
 import com.sun.jersey.api.client.WebResource;
-import eu.ehri.extension.AbstractRestResource;
+import eu.ehri.extension.base.AbstractRestResource;
 import org.junit.Test;
 
 import javax.ws.rs.core.MediaType;
