+++ ehri-ws/src/test/java/eu/ehri/extension/test/GroupRestClientTest.java
@@ -22,7 +22,7 @@ package eu.ehri.extension.test;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.google.common.collect.Sets;
 import com.sun.jersey.api.client.ClientResponse;
-import eu.ehri.extension.AbstractRestResource;
+import eu.ehri.extension.base.AbstractRestResource;
 import eu.ehri.extension.GroupResource;
 import eu.ehri.project.definitions.Entities;
 import eu.ehri.project.persistence.Bundle;
